pyproject.toml
[tool.poetry]
name = "avatar-renderer-pod"
version = "0.1.0"
description = "Synthesia‑class avatar renderer: FOMM + Diff2Lip + Wav2Lip pipeline with MCP STDIO support"
authors = ["Ruslan Magana Vsevolodovna <ruslan@ruslanmv.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
  { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"

# Web framework & server
fastapi = "0.100.0"
uvicorn = { version = "0.23.2", extras = ["standard"] }

# Task queue
celery = "5.4.7"
redis = "4.7.0"

# Core ML libraries
torch = { version = "2.3.0+cu124", extras = ["cuda124"], source = "pytorch" }
torchvision = { version = "0.14.1+cu124", source = "pytorch" }
diffusers = "0.20.2"
transformers = "4.35.0"
safetensors = "0.3.1"

# Face animation & enhancement
gfpgan = "2.3.0"
first-order-model = { git = "https://github.com/AliaksandrSiarohin/first-order-model.git", tag = "v1.0" }
SadTalker = { git = "https://github.com/OpenTalker/SadTalker.git", tag = "v0.1.0" }
Wav2Lip = { git = "https://github.com/Rudrabha/Wav2Lip.git", tag = "v0.1" }

# Networking & cloud
requests = "2.31.0"
boto3 = "1.26.160"
kafka-python = "2.0.2"

# Utilities
python-dotenv = "1.0.0"
pydantic = "2.8.0"

[tool.poetry.dev-dependencies]
pytest = "7.4.0"
black = "24.1.0"
flake8 = "6.1.0"
pre-commit = "3.4.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
default = false
